# Documentation/Swagger.json
{  
   "swagger":"2.0",
   "info":{  
      "description":"Evolvice Restful Endpoints developed by Evolvice team",
      "version":"1.0",
      "title":"Evolvice API",
      "termsOfService":"urn:tos",
      "contact":{  
         "name":"Mostafa Hussien",
         "url":"evolvice.com",
         "email":"mostafa.hussien.moawad@gmail.com"
      },
      "license":{  
         "name":"Apache 2.0",
         "url":"http://www.apache.org/licenses/LICENSE-2.0"
      }
   },
   "host":"localhost:8080",
   "basePath":"/",
...
   "consumes":[  
      "application/xml",
      "application/json"
   ],
   "produces":[  
      "application/xml",
      "application/json"
   ],
   "paths":{  
...
      "/cars":{  
         "get":{  
            "tags":[  
               "car-resource"
            ],
            "summary":"retrieveAllCars",
            "operationId":"retrieveAllCarsUsingGET",
            "consumes":[  
               "application/xml",
               "application/json"
            ],
            "produces":[  
               "application/xml",
               "application/json"
            ],
            "responses":{  
               "200":{  
                  "description":"OK",
                  "schema":{  
                     "type":"array",
                     "items":{  
                        "$ref":"#/definitions/CarDto"
                     }
                  }
               },
               "401":{  
                  "description":"Unauthorized"
               },
               "403":{  
                  "description":"Forbidden"
               },
               "404":{  
                  "description":"Not Found"
               }
            }
         },
         "post":{  
            "tags":[  
               "car-resource"
            ],
            "summary":"addCar",
            "operationId":"addCarUsingPOST",
            "consumes":[  
               "application/xml",
               "application/json"
            ],
            "produces":[  
               "application/xml",
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"body",
                  "name":"carDto",
                  "description":"carDto",
                  "required":true,
                  "schema":{  
                     "$ref":"#/definitions/CarDto"
                  }
               }
            ],
            "responses":{  
               "200":{  
                  "description":"OK",
                  "schema":{  
                     "type":"object"
                  }
               },
               "201":{  
                  "description":"Created"
               },
               "401":{  
                  "description":"Unauthorized"
               },
               "403":{  
                  "description":"Forbidden"
               },
               "404":{  
                  "description":"Not Found"
               }
            }
         }
      },
      "/cars/{id}":{  
         "get":{  
            "tags":[  
               "car-resource"
            ],
            "summary":"retrieveCar",
            "operationId":"retrieveCarUsingGET",
            "consumes":[  
               "application/xml",
               "application/json"
            ],
            "produces":[  
               "application/xml",
               "application/json"
            ],
            "parameters":[  
               {  
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"OK",
                  "schema":{  
                     "$ref":"#/definitions/CarDto"
                  }
               },
               "401":{  
                  "description":"Unauthorized"
               },
               "403":{  
                  "description":"Forbidden"
               },
               "404":{  
                  "description":"Not Found"
               }
            }
         },
         "delete":{  
            "tags":[  
               "car-resource"
            ],
            "summary":"deleteCar",
            "operationId":"deleteCarUsingDELETE",
            "consumes":[  
               "application/xml",
               "application/json"
            ],
            "produces":[  
               "application/xml",
               "application/json"
            ],
            "parameters":[  
               {  
                  "name":"id",
                  "in":"path",
                  "description":"id",
                  "required":true,
                  "type":"integer",
                  "format":"int32"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"OK"
               },
               "401":{  
                  "description":"Unauthorized"
               },
               "204":{  
                  "description":"No Content"
               },
               "403":{  
                  "description":"Forbidden"
               }
            }
         }
      },
...
   },
   "definitions":{  
...
      "CarDto":{  
         "type":"object",
         "properties":{  
            "brand":{  
               "type":"string"
            },
            "color":{  
               "type":"string"
            },
            "id":{  
               "type":"integer",
               "format":"int32"
            },
            "model":{  
               "type":"string"
            },
            "price":{  
               "type":"number",
               "format":"double"
            },
            "yearOfProduction":{  
               "type":"string",
               "format":"date-time",
               "description":"yearOfProduction should be in the past"
            }
         },
         "description":"Details about car model"
      },
...
   }
}